# expensetrackerapi/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Expense Tracker",
    "description": "Expense Tracker",
    "version": "1.0.0"
  },
  "host": "localhost:1337",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/student/retrieve": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "organizationCode",
            "in": "query",
            "type": "string"
          },
          {
            "name": "course",
            "in": "query",
            "type": "string"
          },
          {
            "name": "year",
            "in": "query",
            "type": "string"
          },
          {
            "name": "financialYear",
            "in": "query",
            "type": "string"
          },
          {
            "name": "objectCode",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully."
          }
        }
      }
    },
    "/user/save": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User save service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body.",
            "schema": {
              "$ref": "#/definitions/userSave"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created successfully."
          }
        }
      }
    },
    "/student/viewstudenttransactions": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "student viewstudenttransactions service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "User.",
            "required": true,
            "type": "string"
          },
          {
            "name": "objectCode",
            "in": "query",
            "description": "Object Code.",
            "required": true,
            "type": "string"
          },
          {
            "name": "skipACL",
            "in": "query",
            "description": "skip ACL.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully."
          }
        }
      }
    },
    "/student/viewbyId": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "student viewbyId service",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "User.",
            "required": true,
            "type": "string"
          },
          {
            "name": "objectCode",
            "in": "query",
            "description": "Object Code.",
            "required": true,
            "type": "string"
          },
          {
            "name": "skipACL",
            "in": "query",
            "description": "skip ACL.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}